% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{document_gadget}
\alias{document_gadget}
\title{Launch a gadget to document selected code}
\usage{
document_gadget()
}
\value{
NULL; the gadget is launched in the RStudio viewer pane. The user can
 edit the selected code and save or cancel the changes.
}
\description{
Launch a gadget to document selected code
}
