% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_llm_provider.R
\name{get_llm_provider}
\alias{get_llm_provider}
\title{Get the LLM provider for 'documentWithPrompt'}
\usage{
get_llm_provider()
}
\value{
An object of class 'LlmProvider' (see ?`llm_provider-class` from package 'tidyprompt')
}
\description{
This function retrieves the LLM provider for the 'documentWithPrompt' add-in.
This providers is responsible for handling prompts send to a LLM API.
}
