% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{retrieve_code}
\alias{retrieve_code}
\title{Retrieve the selected code from the source editor}
\usage{
retrieve_code()
}
\value{
A list with the context, range, and code of the selected code
}
\description{
Retrieve the selected code from the source editor
}
\keyword{internal}
