% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_llm_provider.R
\name{set_llm_provider}
\alias{set_llm_provider}
\title{Set the LLM provider for 'documentWithPrompt'}
\usage{
set_llm_provider(llm_provider)
}
\arguments{
\item{llm_provider}{An object of class 'LlmProvider' (see ?`llm_provider-class` from package 'tidyprompt')}
}
\description{
This function sets the LLM provider for the 'documentWithPrompt' add-in.
This providers is responsible for handling prompts send to a LLM API.
}
