% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{document_gadget}
\alias{document_gadget}
\title{Launch a gadget to document code with LLMs}
\usage{
document_gadget()
}
\value{
NULL; this function is called for its side effects (launching the gadget)
}
\description{
Launch a gadget to document code with LLMs
}
