% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_edit_prompt.R
\name{editPromptModuleUI}
\alias{editPromptModuleUI}
\title{Edit Prompt Module UI}
\usage{
editPromptModuleUI(id, label = "Configure prompt")
}
\arguments{
\item{id}{Shiny module ID}

\item{label}{The label for the button that opens the modal}
}
\value{
A tagList containing shiny UI elements
}
\description{
Edit Prompt Module UI
}
