% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_code_selection.R
\name{retrieve_code_selection}
\alias{retrieve_code_selection}
\title{Retrieve the selected code from the source editor}
\usage{
retrieve_code_selection()
}
\value{
A list with the context, range, and code of the selected code
}
\description{
Retrieve the selected code from the source editor
}
\keyword{internal}
